# backend/docs/swagger.yaml
info:
  contact: {}
paths:
...
  /api/cart:
    get:
      consumes:
      - application/json
      description: Retrieves the shopping cart for a specific user using the access
        token stored in cookies.
      parameters:
      - description: Access Token
        in: header
        name: access_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"uid": "string", "userCart": "object"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "string"}'
          schema:
            additionalProperties: true
            type: object
      summary: Get user cart
      tags:
      - cart
  /api/cart/items:
    post:
      consumes:
      - application/json
      description: Adds a product to the user's shopping cart using the access token
        stored in cookies.
      parameters:
      - description: Access Token
        in: header
        name: access_token
        required: true
        type: string
      - description: Product ID and Quantity
        in: body
        name: input
        required: true
        schema:
          properties:
            product_id:
              type: integer
            quantity:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Product successfully added to cart"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "string"}'
          schema:
            additionalProperties: true
            type: object
      summary: Add item to cart
      tags:
      - cart
...
swagger: "2.0"